(#40): 자동차, 보행자 follow분리
