+++ src/pages/map/destination/types/carType.d.ts
@@ -71,6 +71,7 @@ export type TurnType =
   | 187 //세 번째 경유지
   | 188 //네 번째 경유지
   | 189 //다섯 번째 경유지
+  | 121 //터널진입
   | 125 //육교
   | 126 //지하보도
   | 127 //계단 진입
